requirements.txt
appdirs==1.4.4
async-timeout==4.0.2
audioread==3.0.0
autopep8==1.7.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
decorator==5.1.1
Deprecated==1.2.13
dlib==19.24.0
dlib-binary==19.21.99
face-recognition-models==0.3.0
async-timeout==4.0.2
autopep8==1.7.0
Deprecated==1.2.13
gcloud>=0.18.3
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==3.4
imutils==0.5.4
joblib==1.2.0
jws==0.1.3
librosa==0.9.2
llvmlite==0.39.1
nodeenv==1.7.0
numba==0.56.4
numpy==1.23.4
oauth2client>=4.1.2
# opencv-python==4.6.0.66
opencv-python-headless
packaging==21.3
Pillow==9.3.0
pooch==1.6.0
nodeenv==1.7.0
packaging==21.3
protobuf==4.21.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pycryptodome>=3.6.4
pyparsing==3.0.9
pyrebase4==4.5.0
pyright==1.1.279
python-dotenv==0.21.0
python-jwt>=2.0.1
redis==4.3.4
requests>=2.19.1
requests-toolbelt>=0.7.1
resampy==0.4.2
rsa==4.9
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
soundfile==0.11.0
threadpoolctl==3.1.0
toml==0.10.2
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
typing_extensions==4.4.0
urllib3==1.26.12
wrapt==1.14.1
